requirements.txt
# Production Django Backend Requirements - Fixed for Render.com
Django==4.2.16
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
django-environ==0.11.2
psycopg2-binary==2.9.9
dj-database-url==2.1.0
Pillow==10.4.0 --only-binary :all :
celery==5.3.4
redis==5.0.1
django-redis==5.4.0
django-allauth==0.57.0
djangorestframework-simplejwt==5.3.0
stripe==8.0.0
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.6.0
sentry-sdk==1.38.0
django-ratelimit==4.1.0
django-cleanup==8.0.0
django-imagekit==5.0.0
django-taggit==5.0.1
django-mptt==0.15.0
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
cloudinary==1.36.0
django-cloudinary-storage==0.3.0
resend==0.7.0
django-extensions==3.2.3
